Atualizando o projeto aplicando a utilização dos princípios de SOLID, POO e Design Patterns

+++ CadClientes/Model/Cliente.cs
@@ -13,7 +13,7 @@ namespace CadClientes.Model
         public int id { get; set; }
         public string nome { get; set; }
         public DateTime dt_nasc { get; set; }
-        public int cep { get; set; }
+        public long cep { get; set; }
         public string cidade { get; set; }
         public string estado { get; set; }
         public string logradouro { get; set; }
